requirements.txt
click>=8.0
rich>=13.0
httpx>=0.24
pydantic>=2.0
python-dotenv>=1.0
psutil>=5.9
